Completed Assignment
**Reloading:** The current game allows for the player to fire infinitely. We would like you to add
a reloading system. This would allow the designer to specify a *magazine size* and *reload
delay* in the editor. The gun should fire *magazine size* bullets then take *reload delay*
seconds to reload.
**Scoreboard:** The game does not currently keep track of how many enemies you have
downed. Please provide an endgame summary. This scoring should be integrated into the
existing UI.
**Better AI:** The AI currently just sets the player as a target and moves towards the player.
You are tasked with developing different "personalities" for each of the 3 enemy types.
● ZomBunny - should try to move at you while staying out of your line of fire. They will
strafe left or right to move out of the way of the current direction your character is facing.
They will still move towards the player. They also should be easier to kill, and should die
in 2 hits.
● ZomBear - behavior will remain unchanged, however it should die in 4 hits.
● Hellephant - will still move straight at the player. Once it gets within a certain radius it will stop moving for a time and Enrage. We will leave this up to you to determine how “Enraged” visually looks, but mechanically the Enraged Hellephant should speed up.